sqllogictest: use a regular postgres connection
The previous implementation required correctly implementing many parts of
the pgwire protocol (prepare, bind, execute) itself, including dataflow
and coordinator handling. This meant that some SQL features would have
to be re-implemented here in order to be tested (TAIL, cursors). The
initial reason that slt could not use pgwire are now gone, so it is now
reasonable to make this a normal pgwire connector and remove the need
to double implement some things.

This change now means we have stricter type handling in slt files. A
few files have been fixed.

We also format some data (times with microseconds, decimals and floats)
differently in some cases. This is not a change in the actual data
over the wire, just our driver's string display of it (i.e., it doesn't
actually matter). If we have some deep need to support identical output
to third party slt files, that would require plumbing down some more
options to the display functions.

Fixes #2583

+++ src/sqllogictest/src/ast.rs
@@ -115,7 +115,7 @@ pub enum Record<'a> {
     Statement {
         location: Location,
         expected_error: Option<&'a str>,
-        rows_affected: Option<usize>,
+        rows_affected: Option<u64>,
         sql: &'a str,
     },
     /// A `query` directive.
